% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionplot.R
\name{distributionplot}
\alias{distributionplot}
\title{Gene number distribution plot across a given number of genomes in a population.}
\usage{
distributionplot(Csv, Scale, Name)
}
\arguments{
\item{Csv}{IN QUOTES: Relative path to the 'gene_presence_absence.csv' Roary output file.}

\item{Scale}{IN QUOTES: The choosing for continuous ('con') or log ('log') scale. Default as 'con'.}

\item{Name}{IN QUOTES: The name of the saved output plot. The plot will be automatically saved in PNG format thus no need to specify extension.}
}
\value{
A plot showing the distribution of genes present in a given number of genomes in a population
}
\description{
Gene number distribution plot across a given number of genomes in a population.
}
\examples{
distributionplot("testdata/campylobacter.csv","con","test_campy")

}
